requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
bleach==6.1.0
Pillow==10.1.0
